working on city and roadgen algorithms

+++ Geometry/Edge/Edge.cs
@@ -14,7 +14,7 @@ public struct Edge<T>
         T1 = larger(t1, t2) ? t1 : t2;
         T2 = T1.Equals(t1) ? t2 : t1;
     }
-    public Edge(T t1, T t2, Func<T,int> rank)
+    public Edge(T t1, T t2, Func<T,float> rank)
     {
         if (t1.Equals(t2)) throw new Exception();
         T1 = rank(t1) > rank(t2) ? t1 : t2;
